pyproject.toml
[project]
name = "sous-chef"
version = "0.1.0"
description = "YAML-based feature view creation for Feast"
authors = [
    { name = "Baskaran Eswaran", email = "baskaranz@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "feast>=0.42.0",
    "pyyaml>=5.4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sous_chef"]

[project.urls]
Homepage = "https://github.com/baskaranz/sous-chef"
Repository = "https://github.com/baskaranz/sous-chef"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
target-version = "py38"